Revert "Merge pull request #59 from Project-Heimdall/feature-16-adding-pagination-to-the-browse-all-projects"
This reverts commit ac166ce05320f910ac9b61475250774edbfec2ff, reversing
changes made to f2b7795ed38c4a76e0c25ff82ef5a98ccccbad59.

This merge is being reverted because it introduced a paging component and implementation that was subpar.  We need to first how we want paging to work then find the right component that gives us the desired capability.